+++ src/kernel/comstdio.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 void comstrput(const char *s);
-void cbufferprint(const char *message, const void *buffer, uint16_t count);
+void cbufferprint(const char *message, const void *buffer, uint32_t count);
 void cprint(const char *format, ...);
 void comget(char* buffer, char stop);
