+++ src/kernel/memory.c
@@ -2,17 +2,17 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-void* memcpy(void* destination, const void* source, uint16_t number) {
+void* memcpy(void* destination, const void* source, uint32_t number) {
 	uint8_t * unsigned8destination = (uint8_t *)destination;
 	const uint8_t * unsigned8source = (const uint8_t *)source;
 
-	for (uint16_t i = 0; i < number; i++)
+	for (uint32_t i = 0; i < number; i++)
 		unsigned8destination[i] = unsigned8source[i];
 
 	return destination;
 }
 
-void* memset(void* pointer, int value, uint16_t number) {
+void* memset(void* pointer, int value, uint32_t number) {
 	uint8_t* unsigned8pointer = (uint8_t *)pointer;
 
 	for (uint16_t i = 0; i < number; i++)
@@ -21,7 +21,7 @@ void* memset(void* pointer, int value, uint16_t number) {
 	return pointer;
 }
 
-int memcmp(const void* pointer1, const void* pointer2, uint16_t number) {
+int memcmp(const void* pointer1, const void* pointer2, uint32_t number) {
 	const uint8_t* unsigned8pointer1 = (const uint8_t *)pointer1;
 	const uint8_t* unsigned8pointer2 = (const uint8_t *)pointer2;
